MAE-823: Restrict emails in the "to" form field for each case ID
Before now civicrm uses the value in _toContactEmails to know which email to send email to, but it now uses the "to" html field. to restrict the contact emails that will be included in the bulk email for the current form,
we update the emails in the "to" field, removing those that do not belong to the current case id.